Drop test py35 (#2358)
* Merge tests

* Convert all tests

* Remove unused collector

* Fix travis
