+++ .travis.yml
@@ -150,7 +150,7 @@ jobs:
     env:
     - flake8
     script:
-    - flake8 aiohttp --exclude=aiohttp/backport_cookies.py examples tests demos
+    - flake8 aiohttp examples tests demos
 
   - <<: *_lint_base
     env:
