+++ tests/conftest.py
@@ -1,6 +1,5 @@
 import collections
 import logging
-import sys
 
 import pytest
 
@@ -77,9 +76,3 @@ class _AssertLogsContext:
 @pytest.yield_fixture
 def log():
     yield _AssertLogsContext
-
-
-def pytest_ignore_collect(path, config):
-    if 'test_py35' in str(path):
-        if sys.version_info < (3, 5, 0):
-            return True
