+++ tests/test_multipart.py
@@ -1,6 +1,7 @@
 import asyncio
 import functools
 import io
+import json
 import unittest
 import zlib
 from unittest import mock
@@ -1700,3 +1701,72 @@ class ContentDispositionFilenameTestCase(unittest.TestCase):
         params = {'filename': '=?ISO-8859-1?Q?foo-=E4.html?='}
         self.assertEqual('=?ISO-8859-1?Q?foo-=E4.html?=',
                          content_disposition_filename(params))
+
+
+async def test_async_for_reader(loop):
+    data = [
+        {"test": "passed"},
+        42,
+        b'plain text',
+        b'aiohttp\n',
+        b'no epilogue']
+    reader = aiohttp.MultipartReader(
+        headers={CONTENT_TYPE: 'multipart/mixed; boundary=":"'},
+        content=Stream(b'\r\n'.join([
+            b'--:',
+            b'Content-Type: application/json',
+            b'',
+            json.dumps(data[0]).encode(),
+            b'--:',
+            b'Content-Type: application/json',
+            b'',
+            json.dumps(data[1]).encode(),
+            b'--:',
+            b'Content-Type: multipart/related; boundary="::"',
+            b'',
+            b'--::',
+            b'Content-Type: text/plain',
+            b'',
+            data[2],
+            b'--::',
+            b'Content-Disposition: attachment; filename="aiohttp"',
+            b'Content-Type: text/plain',
+            b'Content-Length: 28',
+            b'Content-Encoding: gzip',
+            b'',
+            b'\x1f\x8b\x08\x00\x00\x00\x00\x00\x00\x03K\xcc\xcc\xcf())'
+            b'\xe0\x02\x00\xd6\x90\xe2O\x08\x00\x00\x00',
+            b'--::',
+            b'Content-Type: multipart/related; boundary=":::"',
+            b'',
+            b'--:::',
+            b'Content-Type: text/plain',
+            b'',
+            data[4],
+            b'--:::--',
+            b'--::--',
+            b'',
+            b'--:--',
+            b''])))
+    idata = iter(data)
+
+    async def check(reader):
+        async for part in reader:
+            if isinstance(part, aiohttp.BodyPartReader):
+                if part.headers[CONTENT_TYPE] == 'application/json':
+                    assert next(idata) == (await part.json())
+                else:
+                    assert next(idata) == await part.read(decode=True)
+            else:
+                await check(part)
+
+    await check(reader)
+
+
+async def test_async_for_bodypart(loop):
+    part = aiohttp.BodyPartReader(
+        boundary=b'--:',
+        headers={},
+        content=Stream(b'foobarbaz\r\n--:--'))
+    async for data in part:
+        assert data == b'foobarbaz'
